+++ build.gradle
@@ -12,7 +12,7 @@ configurations {
     provided
 }
 
-version = "0.0.3"
+version = "0.1.0"
 
 sourceCompatibility = 1.7
 targetCompatibility = 1.7
