+++ lib/embulk/input/hdfs.rb
@@ -1,3 +1,3 @@
 Embulk::JavaPlugin.register_input(
-  "hdfs", "org.embulk.input.HdfsFileInputPlugin",
+  "hdfs", "org.embulk.input.hdfs.HdfsFileInputPlugin",
   File.expand_path('../../../../classpath', __FILE__))
